Fix abs() calls
This fixes compiler errors such as:

src/editor.cpp:1425:66: error: call of overloaded ‘abs(size_t)’ is ambiguous
           if (x == gr->getCursorX() && abs(y-(gr->getCursorY()+2)) < 2)
                                                                  ^

+++ src/editor.cpp
@@ -1422,7 +1422,7 @@ bool eventEditor(const SDL_Event & event)
             y = l->getEntryY();
           }
 
-          if (x == gr->getCursorX() && abs(y-(gr->getCursorY()+2)) < 2)
+          if (x == gr->getCursorX() && abs((int)y-(int)(gr->getCursorY()+2)) < 2)
             gr->setStatus(_("Doors must not overlap"));
           else
           {
